// convex/lib/types.ts
export type TransactionType =
  | "purchase"
  | "reward"
  | "sale"
  | "gift"
  | "refund"
  | "conversion"
  | "marketplace_fee"
  | "auction_bid"
  | "auction_refund";